Adds evaluation parameters to metric creation
Adds "evaluation_params" field to the metric creation API.

This field allows users to specify evaluation parameters when the
metric is a Partial Prompt (LLM-as-a-Judge), which will be prepended
to the judge prompt to construct the final prompt.

It should not be provided if the metric is NOT a Partial Prompt.